requirements-otel.txt
# When updating this file, please also bump the versions in
# `requirements-dev.txt`

opentelemetry-distro >= 0.48b0, < 1.0.0
opentelemetry-exporter-otlp >= 1.27.0, < 2.0.0
opentelemetry-instrumentation >= 0.48b0, < 1.0.0
opentelemetry-instrumentation-logging >= 0.48b0, < 1.0.0
